chore: switch lint-staged config to ESM
`.cjs` was unnecessary in a `type: commonjs` project — a plain `.js`
file already would have been CommonJS — and ESM is the better default
for new config files. Behavior unchanged from the previous .cjs
version. Uses `import.meta.dirname` (Node 20.11+); our CI matrix
requires Node 20+.
